[vvasco] added vFilterFlow to compute optical flow using gabor filters - filters are tuned to one velocity, depending on the ratio between temporal and spatial frequency - step of directions set to 0.25 to take into account negative components of the flow

+++ src/processing/vFilterDisparity/include/vFilterDisparity.h
@@ -41,7 +41,7 @@ private:
     int width;
     int pos_x;
     int pos_y;
-    int ts;
+    double ts;
     double theta;
     int fRad;
